+++ Integra/Models/MFX/Equalizer.cs
@@ -22,7 +22,7 @@ namespace Integra.Models.MFX
     /// </remarks>
     public class Equalizer : IToneMFXModel
     {
-        public int Get(int index, int value)
+        public double Get(int index, double value)
         {
             switch (index)
             {
@@ -37,7 +37,7 @@ namespace Integra.Models.MFX
             }
         }
 
-        public int Set(int index, int value)
+        public int Set(int index, double value)
         {
             switch (index)
             {
@@ -65,7 +65,7 @@ namespace Integra.Models.MFX
                     return value.InvalidateRange(0, 127);
 
                 default:
-                    return value;
+                    return (int)value;
             }
         }
     }
